requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
gensim>=4.3.0
glove-python-binary>=0.2.0
tensorflow>=2.13.0
scipy>=1.11.0